% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut_wcm.R
\name{lut_wcm}
\alias{lut_wcm}
\title{Look up table of WCM}
\usage{
lut_wcm(LAI, SM, coeff)
}
\arguments{
\item{LAI}{one dimensional row vector or a range of LAI value}

\item{SM}{one dimensionalrow vector or a range of SM value}

\item{coeff}{Generated A, B, C, D fitted coefficient for WCM using non linear least square using in situ data}
}
\value{
look up table for WCM for given range of LAI and SM
}
\description{
Look up table of WCM
}
\examples{
A= -9.596695
B=-0.005331
C=-11.758309
D=0.011344
lookuptable <- lut_wcm(LAI=seq(1,6,0.1), SM=seq(0,.6,.01),coeff=c(A,B,C,D))
}
